refactor: compat module
