+++ setup.cfg
@@ -7,3 +7,5 @@ exclude =
 per-file-ignores =
     # Need signal handler before imports
     src/ape/__init__.py: E402
+    # Need for imports that seem unused, imported for compat reasons.
+    src/ape/_compat.py: F401
