+++ src/ape/api/networks.py
@@ -4,9 +4,9 @@ from typing import TYPE_CHECKING, Dict, Iterator, List, Optional, Type
 
 from pluggy import PluginManager  # type: ignore
 
+from ape._compat import cached_property
 from ape.exceptions import NetworkError, NetworkNotFoundError
 from ape.types import ABI, AddressType
-from ape.utils import cached_property
 
 from .base import abstractdataclass, abstractmethod, dataclass
 from .config import ConfigItem
