+++ src/ape/managers/compilers.py
@@ -3,12 +3,12 @@ from typing import Dict, List, Set
 
 from dataclassy import dataclass
 
+from ape._compat import cached_property
 from ape.api.compiler import CompilerAPI
 from ape.exceptions import CompilerError
 from ape.logging import logger
 from ape.plugins import PluginManager
 from ape.types import ContractType
-from ape.utils import cached_property
 
 from .config import ConfigManager
