+++ src/ape_test/accounts.py
@@ -3,10 +3,11 @@ from typing import Iterator, List, Optional
 from eth_account import Account as EthAccount  # type: ignore
 from eth_account.messages import SignableMessage
 
+from ape._compat import cached_property
 from ape.api import TestAccountAPI, TestAccountContainerAPI, TransactionAPI
 from ape.convert import to_address
 from ape.types import AddressType, MessageSignature, TransactionSignature
-from ape.utils import GeneratedDevAccount, cached_property, generate_dev_accounts
+from ape.utils import GeneratedDevAccount, generate_dev_accounts
 
 
 class TestAccountContainer(TestAccountContainerAPI):
